# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Order Management API
  version: 1.1.0
  description: API for customer, product, inventory and order-management workflows.
servers:
  - url: http://localhost:3000
paths:
  /health/live:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: Service is alive
  /health/ready:
    get:
      summary: Readiness probe
      responses:
        '200':
          description: Service is ready
  /api/v1/customers:
    post:
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                  minLength: 2
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Customer created
        '409':
          description: Customer email conflict
        '422':
          description: Validation error
  /api/v1/products:
    post:
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sku, name, unitPrice]
              properties:
                sku:
                  type: string
                  minLength: 2
                name:
                  type: string
                  minLength: 2
                unitPrice:
                  type: number
                  exclusiveMinimum: 0
      responses:
        '201':
          description: Product created
        '409':
          description: SKU conflict
        '422':
          description: Validation error
  /api/v1/inventory/{productId}/adjustments:
    post:
      summary: Adjust product inventory
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity:
                  type: integer
      responses:
        '200':
          description: Inventory adjusted
        '404':
          description: Product not found
        '409':
          description: Invalid inventory adjustment
        '422':
          description: Validation error
  /api/v1/orders:
    post:
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerId, items]
              properties:
                customerId:
                  type: string
                  format: uuid
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
      responses:
        '201':
          description: Order created
        '404':
          description: Customer or product not found
        '422':
          description: Validation error
  /api/v1/orders/{orderId}:
    get:
      summary: Get an order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order found
        '404':
          description: Order not found
  /api/v1/orders/{orderId}/confirm:
    post:
      summary: Confirm an order and reserve inventory
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order confirmed
        '404':
          description: Order or inventory not found
        '409':
          description: Invalid state or insufficient inventory
  /api/v1/orders/{orderId}/cancel:
    post:
      summary: Cancel an order and release reserved inventory
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order cancelled
        '404':
          description: Order or inventory not found
        '409':
          description: Invalid order state
  /api/v1/orders/{orderId}/fulfill:
    post:
      summary: Fulfill a confirmed order and consume reserved inventory
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order fulfilled
        '404':
          description: Order not found
        '409':
          description: Only confirmed orders can be fulfilled
        '500':
          description: Reserved inventory invariant violation
components:
  parameters:
    OrderId:
      in: path
      name: orderId
      required: true
      schema:
        type: string
        format: uuid
  headers:
    XRequestId:
      description: Correlation identifier echoed or generated by the service.
      schema:
        type: string