# src/app.ts
import { randomUUID } from "node:crypto";
import express, { type NextFunction, type Request, type Response } from "express";
import { z } from "zod";
import { OrderManagementService } from "./application/order-management-service.js";
import { DomainError } from "./domain/order-model.js";
import { createInMemoryRepositories } from "./infrastructure/in-memory-repositories.js";
import { logger } from "./infrastructure/logger.js";

const customerSchema = z.object({ name: z.string().trim().min(2), email: z.string().trim().email() });
const productSchema = z.object({ sku: z.string().trim().min(2), name: z.string().trim().min(2), unitPrice: z.number().positive() });
const inventorySchema = z.object({ quantity: z.number().int() });
const orderSchema = z.object({
  customerId: z.string().uuid(),
  items: z.array(z.object({ productId: z.string().uuid(), quantity: z.number().int().positive() })).min(1)
});

function createDefaultService() {
  return new OrderManagementService(createInMemoryRepositories());
}

export function createApp(service: OrderManagementService = createDefaultService()) {
  const app = express();
  app.disable("x-powered-by");
  app.use(express.json({ limit: "1mb" }));
  app.use((req, res, next) => {
    const incomingRequestId = req.header("x-request-id")?.trim();
    const requestId = incomingRequestId || randomUUID();
    const startedAt = process.hrtime.bigint();
    res.setHeader("x-request-id", requestId);
    res.on("finish", () => {
      logger.info("http_request_completed", {
        requestId,
        method: req.method,
        path: req.originalUrl,
        statusCode: res.statusCode,
        durationMs: Number(process.hrtime.bigint() - startedAt) / 1_000_000
      });
    });
    next();
  });

  app.get("/health/live", (_req, res) => res.json({ status: "ok" }));
  app.get("/health/ready", (_req, res) => res.json({ status: "ready" }));

  app.post("/api/v1/customers", async (req, res, next) => {
    try {
      const input = customerSchema.parse(req.body);
      res.status(201).json({ data: await service.createCustomer(input) });
    } catch (error) { next(error); }
  });

  app.post("/api/v1/products", async (req, res, next) => {
    try {
      const input = productSchema.parse(req.body);
      res.status(201).json({ data: await service.createProduct(input) });
    } catch (error) { next(error); }
  });

  app.post("/api/v1/inventory/:productId/adjustments", async (req, res, next) => {
    try {
      const { quantity } = inventorySchema.parse(req.body);
      res.json({ data: await service.adjustInventory(req.params.productId, quantity) });
    } catch (error) { next(error); }
  });

  app.post("/api/v1/orders", async (req, res, next) => {
    try {
      const input = orderSchema.parse(req.body);
      res.status(201).json({ data: await service.createOrder(input) });
    } catch (error) { next(error); }
  });

  app.post("/api/v1/orders/:orderId/confirm", async (req, res, next) => {
    try { res.json({ data: await service.confirmOrder(req.params.orderId) }); }
    catch (error) { next(error); }
  });

  app.post("/api/v1/orders/:orderId/cancel", async (req, res, next) => {
    try { res.json({ data: await service.cancelOrder(req.params.orderId) }); }
    catch (error) { next(error); }
  });

  app.post("/api/v1/orders/:orderId/fulfill", async (req, res, next) => {
    try { res.json({ data: await service.fulfillOrder(req.params.orderId) }); }
    catch (error) { next(error); }
  });

  app.get("/api/v1/orders/:orderId", async (req, res, next) => {
    try { res.json({ data: await service.getOrder(req.params.orderId) }); }
    catch (error) { next(error); }
  });

  app.use((error: unknown, req: Request, res: Response, _next: NextFunction) => {
    if (error instanceof z.ZodError) {
      return res.status(422).json({ error: { code: "VALIDATION_ERROR", message: "Request validation failed", details: error.issues } });
    }
    if (error instanceof DomainError) {
      return res.status(error.status).json({ error: { code: error.code, message: error.message } });
    }

    logger.error("http_request_failed", {
      requestId: res.getHeader("x-request-id"),
      method: req.method,
      path: req.originalUrl,
      errorName: error instanceof Error ? error.name : "UnknownError",
      errorMessage: error instanceof Error ? error.message : String(error)
    });
    return res.status(500).json({ error: { code: "INTERNAL_ERROR", message: "Unexpected server error" } });
  });

  return app;
}